+++ Engine/Source/Runtime/RenderCore/IndexBuffer.h
@@ -8,7 +8,7 @@ using Index = uint32_t;
 class IndexBuffer
 {
 public:
-	static std::shared_ptr<IndexBuffer> Create(const uint32_t size, const Index *indices);
+	static std::shared_ptr<IndexBuffer> Create(const size_t size, const Index *indices);
 
 public:
 	virtual ~IndexBuffer() = default;
